+++ libcaf_core/caf/message.cpp
@@ -31,204 +31,112 @@
 namespace caf {
 
 bool message::load(deserializer& source) {
-  // For machine-to-machine data formats, we prefix the type information.
-  if (!source.has_human_readable_format()) {
-    GUARDED(source.begin_object(type_id_v<message>, "message"));
-    GUARDED(source.begin_field("types"));
-    size_t msg_size = 0;
-    GUARDED(source.begin_sequence(msg_size));
-    using uint16_limits = std::numeric_limits<uint16_t>;
-    if (msg_size > static_cast<size_t>(uint16_limits::max() - 1))
-      STOP(sec::invalid_argument, "too many types for message");
-    if (msg_size == 0) {
-      data_.reset();
-      return source.end_sequence()           //
-             && source.end_field()           //
-             && source.begin_field("values") //
-             && source.end_field()           //
-             && source.end_object();
+  GUARDED(source.begin_object(type_id_v<message>, "message"));
+  GUARDED(source.begin_field("types"));
+  size_t msg_size = 0;
+  GUARDED(source.begin_sequence(msg_size));
+  using uint16_limits = std::numeric_limits<uint16_t>;
+  if (msg_size > static_cast<size_t>(uint16_limits::max() - 1))
+    STOP(sec::invalid_argument, "too many types for message");
+  if (msg_size == 0) {
+    data_.reset();
+    return source.end_sequence()           //
+           && source.end_field()           //
+           && source.begin_field("values") //
+           && source.begin_tuple(0)        //
+           && source.end_tuple()           //
+           && source.end_field()           //
+           && source.end_object();
+  }
+  detail::type_id_list_builder ids{msg_size};
+  size_t data_size = 0;
+  if (source.has_human_readable_format()) {
+    for (size_t i = 0; i < msg_size; ++i) {
+      std::string type_name;
+      GUARDED(source.value(type_name));
+      auto type = source.resolve_type_id(type_name);
+      if (auto meta = detail::global_meta_object_or_null(type))
+        data_size += meta->padded_size;
+      else
+        STOP(sec::unknown_type, type_name);
+      ids.push_back(type);
     }
-    detail::type_id_list_builder ids{msg_size};
-    size_t data_size = 0;
+  } else {
     for (size_t i = 0; i < msg_size; ++i) {
       type_id_t id = 0;
       GUARDED(source.value(id));
       if (auto meta = detail::global_meta_object_or_null(id))
         data_size += meta->padded_size;
       else
-        STOP(sec::unknown_type);
+        STOP(sec::unknown_type, id);
       ids.push_back(id);
     }
-    GUARDED(source.end_sequence());
-    CAF_ASSERT(ids.size() == msg_size);
-    intrusive_ptr<detail::message_data> ptr;
-    if (auto vptr = malloc(sizeof(detail::message_data) + data_size)) {
-      // We don't need to worry about exceptions here: the message_data
-      // constructor as well as `move_to_list` are `noexcept`.
-      ptr.reset(new (vptr) detail::message_data(ids.move_to_list()), adopt_ref);
-    } else {
-      STOP(sec::runtime_error, "unable to allocate memory");
-    }
-    auto pos = ptr->storage();
-    auto types = ptr->types();
-    auto gmos = detail::global_meta_objects();
-    GUARDED(source.begin_field("values"));
-    GUARDED(source.begin_tuple(msg_size));
-    for (size_t i = 0; i < msg_size; ++i) {
-      auto& meta = gmos[types[i]];
-      meta.default_construct(pos);
-      ptr->inc_constructed_elements();
-      if (!meta.load(source, pos))
-        return false;
-      pos += meta.padded_size;
-    }
-    data_.reset(ptr.release(), adopt_ref);
-    return source.end_tuple() && source.end_field() && source.end_object();
   }
-  // For human-readable data formats, we serialize messages as a single list of
-  // dynamically-typed objects. This is more expensive, because we first need
-  // the full type information before we can allocate the storage. Hence, we
-  // must deserialize each element into a separate memory block first and then
-  // move them to their final destination later.
-  struct object_ptr {
-    void* obj;
-    const detail::meta_object* meta;
-    object_ptr(void* obj, const detail::meta_object* meta) noexcept
-      : obj(obj), meta(meta) {
-      // nop
-    }
-    object_ptr(object_ptr&& other) noexcept : obj(nullptr), meta(nullptr) {
-      using std::swap;
-      swap(obj, other.obj);
-      swap(meta, other.meta);
-    }
-    object_ptr& operator=(object_ptr&& other) noexcept {
-      if (this != &other) {
-        if (obj) {
-          meta->destroy(obj);
-          free(obj);
-          obj = nullptr;
-          meta = nullptr;
-        }
-        using std::swap;
-        swap(obj, other.obj);
-        swap(meta, other.meta);
-      }
-      return *this;
-    }
-    ~object_ptr() noexcept {
-      if (obj) {
-        meta->destroy(obj);
-        free(obj);
-      }
-    }
-  };
-  struct free_t {
-    void operator()(void* ptr) const noexcept {
-      return free(ptr);
-    }
-  };
-  using unique_void_ptr = std::unique_ptr<void, free_t>;
-  auto msg_size = size_t{0};
-  GUARDED(source.begin_sequence(msg_size));
-  if (msg_size > 0) {
-    // Deserialize message elements individually.
-    std::vector<object_ptr> objects;
-    detail::type_id_list_builder ids;
-    objects.reserve(msg_size);
-    auto data_size = size_t{0};
-    for (size_t i = 0; i < msg_size; ++i) {
-      auto type = type_id_t{0};
-      GUARDED(source.fetch_next_object_type(type));
-      if (auto meta_obj = detail::global_meta_object_or_null(type)) {
-        ids.push_back(type);
-        auto obj_size = meta_obj->padded_size;
-        data_size += obj_size;
-        unique_void_ptr vptr{malloc(obj_size)};
-        if (vptr == nullptr)
-          STOP(sec::runtime_error, "unable to allocate memory");
-        meta_obj->default_construct(vptr.get());
-        objects.emplace_back(vptr.release(), meta_obj);
-        if (!meta_obj->load(source, objects.back().obj))
-          return false;
-      } else {
-        STOP(sec::unknown_type);
-      }
-    }
-    GUARDED(source.end_sequence());
-    // Merge elements into a single message data object.
-    intrusive_ptr<detail::message_data> ptr;
-    if (auto vptr = malloc(sizeof(detail::message_data) + data_size)) {
-      // We don't need to worry about exceptions here: the message_data
-      // constructor as well as `move_to_list` are `noexcept`.
-      ptr.reset(new (vptr) detail::message_data(ids.move_to_list()), adopt_ref);
-    } else {
-      STOP(sec::runtime_error, "unable to allocate memory");
-    }
-    auto pos = ptr->storage();
-    for (auto& x : objects) {
-      // TODO: avoid extra copy by adding move_construct to meta objects
-      x.meta->copy_construct(pos, x.obj);
-      ptr->inc_constructed_elements();
-      pos += x.meta->padded_size;
-    }
-    data_.reset(ptr.release(), adopt_ref);
-    return true;
+  GUARDED(source.end_sequence());
+  GUARDED(source.end_field());
+  CAF_ASSERT(ids.size() == msg_size);
+  intrusive_ptr<detail::message_data> ptr;
+  if (auto vptr = malloc(sizeof(detail::message_data) + data_size)) {
+    // We don't need to worry about exceptions here: the message_data
+    // constructor as well as `move_to_list` are `noexcept`.
+    ptr.reset(new (vptr) detail::message_data(ids.move_to_list()), adopt_ref);
   } else {
-    data_.reset();
-    return source.end_sequence();
+    STOP(sec::runtime_error, "unable to allocate memory");
   }
+  auto pos = ptr->storage();
+  auto types = ptr->types();
+  auto gmos = detail::global_meta_objects();
+  GUARDED(source.begin_field("values"));
+  GUARDED(source.begin_tuple(msg_size));
+  for (size_t i = 0; i < msg_size; ++i) {
+    auto& meta = gmos[types[i]];
+    meta.default_construct(pos);
+    ptr->inc_constructed_elements();
+    if (!meta.load(source, pos))
+      return false;
+    pos += meta.padded_size;
+  }
+  data_.reset(ptr.release(), adopt_ref);
+  return source.end_tuple() && source.end_field() && source.end_object();
 }
 
 bool message::save(serializer& sink) const {
   auto gmos = detail::global_meta_objects();
-  // For machine-to-machine data formats, we prefix the type information.
-  if (!sink.has_human_readable_format()) {
-    if (data_ == nullptr) {
-      // Short-circuit empty tuples.
-      return sink.begin_object(type_id_v<message>, "message") //
-             && sink.begin_field("types")                     //
-             && sink.begin_sequence(0)                        //
-             && sink.end_sequence()                           //
-             && sink.end_field()                              //
-             && sink.begin_field("values")                    //
-             && sink.begin_tuple(0)                           //
-             && sink.end_tuple()                              //
-             && sink.end_field()                              //
-             && sink.end_object();
-    }
-    GUARDED(sink.begin_object(type_id_v<message>, "message"));
-    auto type_ids = data_->types();
-    // Write type information.
-    GUARDED(sink.begin_field("types") && sink.begin_sequence(type_ids.size()));
-    for (auto id : type_ids)
-      GUARDED(sink.value(id));
-    GUARDED(sink.end_sequence() && sink.end_field());
-    // Write elements.
-    auto storage = data_->storage();
-    GUARDED(sink.begin_field("values") && sink.begin_tuple(type_ids.size()));
-    for (auto id : type_ids) {
-      auto& meta = gmos[id];
-      GUARDED(meta.save(sink, storage));
-      storage += meta.padded_size;
-    }
-    return sink.end_tuple() && sink.end_field() && sink.end_object();
-  }
-  // For human-readable data formats, we serialize messages as a single list of
-  // dynamically-typed objects.
   if (data_ == nullptr) {
     // Short-circuit empty tuples.
-    return sink.begin_sequence(0) && sink.end_sequence();
+    return sink.begin_object(type_id_v<message>, "message") //
+           && sink.begin_field("types")                     //
+           && sink.begin_sequence(0)                        //
+           && sink.end_sequence()                           //
+           && sink.end_field()                              //
+           && sink.begin_field("values")                    //
+           && sink.begin_tuple(0)                           //
+           && sink.end_tuple()                              //
+           && sink.end_field()                              //
+           && sink.end_object();
   }
+  GUARDED(sink.begin_object(type_id_v<message>, "message"));
   auto type_ids = data_->types();
-  GUARDED(sink.begin_sequence(type_ids.size()));
+  // Write type information. This is needed for serialization round trips. Human
+  // readable formats will have human readable type information.
+  GUARDED(sink.begin_field("types") && sink.begin_sequence(type_ids.size()));
+  if (sink.has_human_readable_format()) {
+    for (auto id : type_ids)
+      GUARDED(sink.value(sink.resolve_type_name(id)));
+  } else {
+    for (auto id : type_ids)
+      GUARDED(sink.value(id));
+  }
+  GUARDED(sink.end_sequence() && sink.end_field());
+  // Write elements.
   auto storage = data_->storage();
+  GUARDED(sink.begin_field("values") && sink.begin_tuple(type_ids.size()));
   for (auto id : type_ids) {
     auto& meta = gmos[id];
     GUARDED(meta.save(sink, storage));
     storage += meta.padded_size;
   }
-  return sink.end_sequence();
+  return sink.end_tuple() && sink.end_field() && sink.end_object();
 }
 
 bool message::save(detail::stringification_inspector& sink) const {